Task updates
> Added task manager
>> Removed task parts from player manager to task manager

> Player now returns game manager reference

> Player now knows which player they are
