+++ Assets/Prefabs/Player.prefab
@@ -83,8 +83,8 @@ Transform:
   - {fileID: 3702365987019445124}
   - {fileID: 3643417929718204416}
   - {fileID: 5413478257145270146}
-  - {fileID: 1481701818224543020}
   - {fileID: 6024997155386976153}
+  - {fileID: 1481701818224543020}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -317,7 +317,7 @@ RectTransform:
   m_Children:
   - {fileID: 2494637358676408422}
   m_Father: {fileID: 4875622312694227661}
-  m_RootOrder: 5
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -692,7 +692,7 @@ RectTransform:
   m_Children:
   - {fileID: 3193049181367931837}
   m_Father: {fileID: 4875622312694227661}
-  m_RootOrder: 6
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -1000,7 +1000,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5728035041552772441, guid: c81e032e02b8b674eaa608a69ccdb343, type: 3}
       propertyPath: m_AnchoredPosition.y
-      value: -0.00021362305
+      value: -0.00024414062
       objectReference: {fileID: 0}
     - target: {fileID: 6317720234434829636, guid: c81e032e02b8b674eaa608a69ccdb343, type: 3}
       propertyPath: m_Camera
