+++ Assets/Scripts/Player/PlayerControls.cs
@@ -136,7 +136,7 @@ namespace ILOVEYOU
                 m_health -= damage;
                 m_iframesCurrent = m_iframesTotal;
                 UpdateHealthBar();
-                if (m_health <= 0) m_plaMa.GetGameManager().PlayerDeath(m_plaMa);
+                if (m_health <= 0) m_plaMa.GetGameManager.PlayerDeath(m_plaMa);
                 
             }
