+++ Assets/Scripts/Player/PlayerManager.cs
@@ -1,9 +1,5 @@
 using ILOVEYOU.Cards;
 using ILOVEYOU.Management;
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Net.Http.Headers;
 using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.UI;
@@ -12,175 +8,52 @@ namespace ILOVEYOU
 {
     namespace Player
     {
+        [RequireComponent(typeof(TaskManager))]
         public class PlayerManager : MonoBehaviour
         {
             [SerializeField] private bool m_debugging;
-            private List<Task> m_tasks = new List<Task>();
-            public int NumberOfTasks { get { return m_tasks.Count; } }
-            [HideInInspector] public int TaskCompletionPoints;
+            //player
+            private int m_playerID;
+            public int GetPlayerID { get { return m_playerID; } }
+            private GameManager m_manager;
+            public GameManager GetGameManager { get { return m_manager; } }
+            //tasks
+            private TaskManager m_taskMan;
+            public TaskManager GetTaskManager { get { return m_taskMan; } }
+            //cards
             private DisruptCard[] m_cardsHeld;
             public bool CardsInHand { get { return m_cardsHeld.Length > 0; } }
             [SerializeField] private float m_cardTimeout;
             [SerializeField] private GameObject m_blindBox;
             //ui
-            //[SerializeField] private GameObject m_taskUIElementPrefab;
-            //private List<GameObject> m_taskUIElements = new List<GameObject>();
-            //[SerializeField] private Transform m_taskUIContainer;
             [SerializeField] private Transform m_cardDisplay;
             [SerializeField] private GameObject m_playerHud;
             private Slider m_healthSlider;
-            private GameManager m_manager;
             public bool Startup(GameManager manager, int playerNum)
             {
                 //Reset variables
-                TaskCompletionPoints = 0;
+                m_taskMan = GetComponent<TaskManager>();
+                if (!m_taskMan)
+                {
+                    if (m_debugging) Debug.Log("Task manager not found! Please fix boss");
+                    return false;
+                }
+                if (!m_taskMan.Startup())
+                {
+                    return false;
+                }
+
                 m_cardsHeld = new DisruptCard[0];
-                if(m_debugging) Debug.Log("PlayerManager started successfully");
                 m_manager = manager;
-                if (playerNum != 0) m_playerHud.transform.GetChild(0).localScale = new(-1, 1, 1);
+                m_playerID = playerNum;
+
+                //ui setup
+                if (m_playerID != 0) m_playerHud.transform.GetChild(0).localScale = new(-1, 1, 1);
                 m_healthSlider = m_playerHud.transform.GetChild(0).GetComponentInChildren<Slider>();
-                return true;
-            }
-            public GameManager GetGameManager() { return m_manager; }
-            #region Task Management
-            /// <summary>
-            /// Creates a new class to add to the list
-            /// </summary>
-            /// <param name="type"></param>
-            /// <param name="cap"></param>
-            /// <returns></returns>
-            public bool AddTask(TaskType type, float cap)
-            {
-                m_tasks.Add(new Task(type, cap));
-                //GameObject taskUI = Instantiate(m_taskUIElementPrefab);
-                //m_taskUIElements.Add(taskUI);
-                //taskUI.transform.SetParent(m_taskUIContainer, false);
-                VerifyTaskList();
-                return true;
-            }
-            public bool AddTask(Task task)
-            {
-                AddTask(task.GetTaskType, task.GetCapValue);
-                return true;
-            }
-            /// <summary>
-            /// Checks if any tasks are complete and removes them from the list
-            /// </summary>
-            /// <returns></returns>
-            private bool VerifyTaskList()
-            {
-                List<Task> completeTasks = new List<Task>();
-                //Get a list of all the completed tasks
-                foreach(Task task in m_tasks)
-                {
-                    if (task.IsComplete)
-                        completeTasks.Add(task);
-                }
-                //Remove those completed tasks from the main list
-                foreach(Task task in completeTasks)
-                {
-                    //Destroy(m_taskUIElements[m_tasks.IndexOf(task)]);
-                    m_tasks.Remove(task);
-                    TaskCompletionPoints++;
-                }
-                return true;
-            }
-            /// <summary>
-            /// Returns all the indexes of tasks matching the given type
-            /// </summary>
-            /// <param name="type"></param>
-            /// <returns></returns>
-            public int[] GetMatchingTasks(TaskType type)
-            {
-                //Create a list of indexes
-                List<int> indexes = new List<int>();
-                //If the type of the current iteration matches the requested type...
-                //...saved its index
-                for (int i = 0; i < m_tasks.Count; i++)
-                {
-                    if (m_tasks[i].GetTaskType == type)
-                    {
-                        indexes.Add(i);
-                    }
-                }
-                return indexes.ToArray();
-            }
-            /// <summary>
-            /// Updates the value of a task at a given index
-            /// </summary>
-            /// <param name="index"></param>
-            /// <param name="value"></param>
-            /// <returns></returns>
-            public bool UpdateTask(int index, float value)
-            {
-                bool x = m_tasks[index].UpdateTask(value);
-                VerifyTaskList();
-                return x;
-            }
-            /// <summary>
-            /// Resets the task at the given index
-            /// </summary>
-            /// <param name="index"></param>
-            /// <returns></returns>
-            public bool ResetTask(int index)
-            {
-                bool x = m_tasks[index].ResetTask();
-                VerifyTaskList();
-                return x;
-            }
-            /// <summary>
-            /// Updates any kill type tasks
-            /// </summary>
-            /// <param name="value"></param>
-            /// <returns></returns>
-            public bool UpdateKillTrackers(float value)
-            {
-                foreach (Task task in m_tasks)
-                {
-                    if (task.GetTaskType == TaskType.Kills)
-                        task.UpdateTask(value);
-                }
-                VerifyTaskList();
-                return true;
-            }
 
-            /// <summary>
-            /// Updates any time type tasks
-            /// </summary>
-            /// <param name="doReset"></param>
-            /// <returns></returns>
-            public bool UpdateTimers(bool doReset)
-            {
-                foreach (Task task in m_tasks)
-                {
-                    if (task.GetTaskType == TaskType.Time)
-                    {
-                        if (doReset)
-                            task.ResetTask();
-                        else
-                            task.UpdateTask(Time.deltaTime);
-                    }
-                }
-                VerifyTaskList();
+                if(m_debugging) Debug.Log("PlayerManager started successfully");
                 return true;
             }
-            /*private void Update()
-            {
-                UpdateTaskUI();
-            }
-            private void UpdateTaskUI()
-            {
-                for (int i = 0; i < m_tasks.Count; i++)
-                {
-                    Image tmp = m_taskUIElements[i].GetComponentInChildren<Image>();
-                    if (tmp != null)
-                    {
-                    tmp.fillAmount = m_tasks[i].GetPercent;
-
-                    }
-                }
-            }*/
-            #endregion
             #region Card Management
             /// <summary>
             /// Save a given hand as the cards held by this player
@@ -189,15 +62,17 @@ namespace ILOVEYOU
             public void CollectHand(DisruptCard[] cards)
             {
                 if(m_debugging) Debug.Log("Hand dealt, setting up cards.");
+
                 //Copy the given array to this hand
                 m_cardsHeld = new DisruptCard[cards.Length];
                 cards.CopyTo(m_cardsHeld, 0);
+
                 //Set up each card
                 foreach(DisruptCard card in m_cardsHeld)
                 {
-                    if(m_debugging) Debug.Log("Readying discard function to card.");
                     card.transform.SetParent(m_cardDisplay, false);
                     card.transform.localScale = Vector3.one;
+                    if(m_debugging) Debug.Log("Readying discard function to card.");
                     card.m_playerHandToDiscard.AddListener(delegate { DiscardHand(); });
                 }
                 //To stop stockpiling, delete the cards after a set time
@@ -223,6 +98,7 @@ namespace ILOVEYOU
                 //If the hand is empty, don't continue
                 if (!CardsInHand)
                     return;
+
                 //Get the vector of the face buttons
                 Vector2 selection = value.Get<Vector2>();
                 if(m_debugging) Debug.Log($"The inputed value {selection}");
@@ -243,17 +119,19 @@ namespace ILOVEYOU
                         index = 2;
                         break;
                 }
-                //Trigger the effects of the chosen card.
+
+                //Trigger the effects of the chosen card if a valid input was given.
                 if(index > - 1)
                 m_cardsHeld[index].Trigger(m_manager, this);
             }
             #endregion
             public void TriggerBlindness(float m_time)
             {
+                CancelInvoke();
                 m_blindBox.SetActive(true);
-                Invoke("DisableBlindness", m_time);
+                Invoke("_disableBlindness", m_time);
             }
-            private void DisableBlindness()
+            private void _disableBlindness()
             {
                 m_blindBox.SetActive(false);
             }
