+++ Assets/Scripts/Player/TaskManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 397e39359dc44dd4baa601e48113eb7e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
